+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-* 0.1.0 (2023-10-31): Initial release
-* 0.1.1 (2023-10-31): Add a function to clean the text (e.g. replace `&nbsp;` with a space, etc);
+* 0.1.0 (2023-10-13): Initial release
+* 0.1.1 (2023-10-13): Add a function to clean the text (e.g. replace `&nbsp;` with a space, etc);
   return error instead of panic
+* 0.1.2 (2023-10-13): Rename `Element.role_ids` to `Element.roles`; rename `pretty.json` to
+  `data-pretty.json`; add `--extended` option; fix changelog dates